put rental function in new class and fix linter errors
